third_party/blink/web_tests/external/wpt/editing/run/createlink-expected.txt

This is a testharness.js-based test.
Found 15 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] [["createlink","http://www.google.com/"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foobar" but got "foo<a href=\\"http://www.google.com/\\">http://www.google.com/</a>bar"
[FAIL] [["createlink","http://www.google.com/"]] "<p>[foo</p><p> <span>bar</span> </p><p>baz]</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><a href=\\"http://www.google.com/\\">foo</a></p><p> <a href=\\"http://www.google.com/\\"><span>bar</span></a> </p><p><a href=\\"http://www.google.com/\\">baz</a></p>" but got "<p><a href=\\"http://www.google.com/\\">foo</a></p><p> <a href=\\"http://www.google.com/\\"><span>bar</span> </a></p><p><a href=\\"http://www.google.com/\\">baz</a></p>"
[FAIL] [["createlink","http://www.google.com/"]] "<b>foo[]bar</b>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<b>foobar</b>" but got "<b>foo<a href=\\"http://www.google.com/\\">http://www.google.com/</a>bar</b>"
[FAIL] [["createlink","http://www.google.com/"]] "<i>foo[]bar</i>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<i>foobar</i>" but got "<i>foo<a href=\\"http://www.google.com/\\">http://www.google.com/</a>bar</i>"
[FAIL] [["createlink","http://www.google.com/"]] "<span>foo</span>{}<span>bar</span>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span>foo</span><span>bar</span>" but got "<span>foo<a href=\\"http://www.google.com/\\">http://www.google.com/</a></span><span>bar</span>"
[FAIL] [["createlink","http://www.google.com/"]] "<span>foo[</span><span>]bar</span>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span>foo</span><span>bar</span>" but got "<span>foo<a href=\\"http://www.google.com/\\">http://www.google.com/</a></span><span>bar</span>"
[FAIL] [["createlink","http://www.google.com/"]] "<a href=otherurl>foo[bar]baz</a>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<a href=\\"http://www.google.com/\\">foobarbaz</a>" but got "<a href=\\"otherurl\\">foo</a><a href=\\"http://www.google.com/\\">bar</a><a href=\\"otherurl\\">baz</a>"
[FAIL] [["createlink","http://www.google.com/"]] "<a href=otherurl>foo[barbaz</a>}" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<a href=\\"http://www.google.com/\\">foobarbaz</a>" but got "<a href=\\"otherurl\\">foo</a><a href=\\"http://www.google.com/\\">barbaz</a>"
[FAIL] [["createlink","http://www.google.com/"]] "{<a href=otherurl>foobar]baz</a>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<a href=\\"http://www.google.com/\\">foobarbaz</a>" but got "<a href=\\"http://www.google.com/\\">foobar</a><a href=\\"otherurl\\">baz</a>"
[FAIL] [["createlink","http://www.google.com/"]] "<a href=otherurl><b>foo[bar]baz</b></a>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<a href=\\"http://www.google.com/\\"><b>foobarbaz</b></a>" but got "<b><a href=\\"otherurl\\">foo</a><a href=\\"http://www.google.com/\\">bar</a><a href=\\"otherurl\\">baz</a></b>"
[FAIL] [["createlink","http://www.google.com/"]] "<a href=otherurl><b>foo[barbaz</b></a>}" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<a href=\\"http://www.google.com/\\"><b>foobarbaz</b></a>" but got "<b><a href=\\"otherurl\\">foo</a><a href=\\"http://www.google.com/\\">barbaz</a></b>"
[FAIL] [["createlink","http://www.google.com/"]] "{<a href=otherurl><b>foobar]baz</b></a>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<a href=\\"http://www.google.com/\\"><b>foobarbaz</b></a>" but got "<b><a href=\\"http://www.google.com/\\">foobar</a><a href=\\"otherurl\\">baz</a></b>"
[FAIL] [["createlink","http://www.google.com/"]] "<a href=otherurl><b>[foobarbaz]</b></a>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<a href=\\"http://www.google.com/\\"><b>foobarbaz</b></a>" but got "<b><a href=\\"http://www.google.com/\\">foobarbaz</a></b>"
[FAIL] [["createlink","http://www.google.com/"]] "<a name=abc>foo[bar]baz</a>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span name=\\"abc\\">foo<a href=\\"http://www.google.com/\\">bar</a>baz</span>" but got "<a name=\\"abc\\">foo</a><a href=\\"http://www.google.com/\\">bar</a><a name=\\"abc\\">baz</a>"
[FAIL] [["createlink","http://www.google.com/"]] "<a name=abc><b>foo[bar]baz</b></a>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span name=\\"abc\\"><b>foo<a href=\\"http://www.google.com/\\">bar</a>baz</b></span>" but got "<b><a name=\\"abc\\">foo</a><a href=\\"http://www.google.com/\\">bar</a><a>baz</a></b>"
Harness: the test ran to completion.
